+++ docs/pages/docs/_meta.json
@@ -6,6 +6,7 @@
   "styling-theming": "Styling & Theming",
   "ui-components": "UI Components",
   "custom-schemas": "Custom Schemas",
+  "collaboration": "Collaboration",
   "advanced": "Advanced",
   "discord-link": {
     "title": "Community ↗",
