+++ docs/pages/docs/collaboration.mdx
@@ -1,0 +1,11 @@
+---
+title: Collaboration
+description: Learn how to create multiplayer experiences with BlockNote
+---
+
+# Collaboration (advanced)
+
+BlockNote supports multi-user collaborative document editing.
+
+- [Real-time collaboration](/docs/collaboration/real-time-collaboration)
+- [Comments](/docs/collaboration/comments)
