+++ docs/pages/docs/collaboration/_meta.json
@@ -1,0 +1,4 @@
+{
+  "real-time-collaboration": "Real-time collaboration",
+  "comments": "Comments"
+}
